move homebrew parsing stuff into homebrew model
