+++ cogs5e/funcs/lookupFuncs.py
@@ -13,7 +13,6 @@ import newrelic.agent
 
 from cogs5e.models.background import Background
 from cogs5e.models.errors import NoActiveBrew
-from cogs5e.models.homebrew.bestiary import Bestiary
 from cogs5e.models.homebrew import Tome
 from cogs5e.models.monster import Monster
 from cogs5e.models.race import Race
@@ -188,6 +187,9 @@ def get_homebrew_formatted_name(named):
 
 
 # ----- Monster stuff
+from cogs5e.models.homebrew.bestiary import Bestiary  # no circular imports here, nope
+
+
 async def select_monster_full(ctx, name, cutoff=5, return_key=False, pm=False, message=None, list_filter=None,
                               return_metadata=False, extra_choices=None, selectkey=None):
     """
